Extract fidget-solver
