+++ Cargo.toml
@@ -7,6 +7,7 @@ members = [
     "fidget-rhai",
     "fidget-mesh",
     "fidget-jit",
+    "fidget-solver",
 
     # Non-WASM demos
     "demos/constraints",
@@ -44,6 +45,7 @@ opt-level = 0
 
 [workspace.dependencies]
 anyhow = "1"
+approx = "0.5.1"
 arrayvec = "0.7"
 bincode = "1.3"
 clap = { version = "4", features = ["derive"] }
