+++ fidget-core/src/lib.rs
@@ -7,4 +7,3 @@ pub use error::Error;
 pub mod gui;
 pub mod render;
 pub mod shapes;
-pub mod solver;
